Moved activate to utils
