+++ matchmaking/backend/controllers/auth.js
@@ -1,8 +1,6 @@
 const { FRONTEND_URL } = require("../config");
-const { keycloak } = require("../utils");
-const { usersQueries } = require("../queries");
-const { getAccessToken, getAuthorizationUrl, getLogoutUrl, getUserData } =
-  keycloak;
+const { keycloak, activateUser } = require("../utils");
+const { getAccessToken, getAuthorizationUrl, getLogoutUrl } = keycloak;
 
 /**
  * Prompts the user to login.
@@ -38,41 +36,13 @@ exports.callback = async (req, res) => {
     const tokens = await getAccessToken({ code, baseURL });
     const redirectUrl = new URL(FRONTEND_URL);
     redirectUrl.searchParams.set("token", tokens.access_token);
-    await activate(tokens.access_token);
+    await activateUser(tokens.access_token);
     res
       .cookie("refresh_token", tokens.refresh_token, { httpOnly: true })
       .redirect(redirectUrl);
   } catch (error) {
     console.error("Controller: Error in login callback", error);
     res.json({ success: false, error: error.message || error });
-  }
-};
-
-/**
- * Adds user to database.
- * @author Brady Mitchell <braden.mitchell@gov.bc.ca | braden.jr.mitch@gmail.com>
- */
-const activate = async (access_token) => {
-  try {
-    const user = getUserData(access_token);
-    if (!user) console.error("Activate: User not found.");
-    else {
-      // Get user from database matching the user guid on the keycloak jwt.
-      const dbUser = await usersQueries.getUserById(user.idir_user_guid)[0];
-
-      if (dbUser === undefined) {
-        // No such user exists, create new user.
-        const createdUser = await usersQueries.createUser(
-          user.idir_user_guid,
-          null,
-          null
-        )[0];
-
-        if (!createdUser) console.error("Activate: User could not be created.");
-      }
-    }
-  } catch (error) {
-    console.error("Controller: Error in activate", error);
   }
 };
