+++ matchmaking/backend/utils/index.js
@@ -1,3 +1,4 @@
 // Util exports
 exports.colors = require("./colors");
 exports.keycloak = require("./keycloak");
+exports.activateUser = require("./activateUser");
